# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Abroad Matrimony API
  version: 1.0.0
  description: |
    REST API for the Abroad Matrimony matchmaking platform.

    **Authentication:** All user endpoints (except OTP request/verify and token refresh)
    require a Bearer access token issued by `POST /api/v1/auth/otp/verify` or
    `POST /api/v1/auth/token/refresh`.

    Admin endpoints require a Bearer admin token from `POST /admin/auth/login`.

    Webhook endpoints (`/stripe/webhook`, `/razorpay/webhook`) use provider signature
    verification instead of Bearer tokens.

    **Response envelope:** Every response follows `ApiResponse<T>`:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "hasMore": false, "cursor": null },
      "requestId": "req_abc123"
    }
    ```
    Errors add `"error": { "code": "ERROR_CODE", "message": "Human-readable message" }`.

    **All money values** are integers in paise (INR) or cents (USD). Never floats.

  contact:
    name: Abroad Matrimony Engineering
  license:
    name: Private
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.abroadmatrimony.com
    description: Production

# ── Security Schemes ──────────────────────────────────────────────────────────
components:
  securitySchemes:
    userBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token issued by /api/v1/auth/otp/verify or /token/refresh. Expires in 15 minutes.
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT issued by /admin/auth/login. Longer-lived; carries AdminRole claim.

# ── Reusable Schemas ──────────────────────────────────────────────────────────
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:    { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: "Phone must be in E.164 format" }
        details: { type: object, additionalProperties: true }
...
    FlagReason:
      type: string
      enum: [INAPPROPRIATE_CONTENT, HARASSMENT, SPAM, FAKE_IDENTITY, UNDERAGE_CONCERN, OTHER]
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:   { type: boolean, enum: [false] }
        error:     { $ref: '#/components/schemas/ApiError' }
        requestId: { type: string }

  # ── Reusable Parameters ───────────────────────────────────────────────────
...
  responses:
    400BadRequest:
      description: Validation error or bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            error: { code: VALIDATION_ERROR, message: "Phone must be in E.164 format" }
    401Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            success: false
            error: { code: UNAUTHORIZED, message: "Access token is invalid or expired" }
...
    404NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    500InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }

# ── Tags ──────────────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/trust/block:
    post:
      tags: [Trust Center]
      summary: Block a user
      description: Cancels any pending connection between the two users.
      operationId: blockUser
      security:
        - userBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId]
              properties:
                userId: { type: string, format: uuid }
                reason: { type: string, maxLength: 500 }
            example:
              userId: "usr-uuid-arjun"
      responses:
        '200':
          description: User blocked
          content:
            application/json:
              example:
                success: true
                data: { id: "block-uuid", blockedUserId: "usr-uuid-arjun", blockedUserName: "Arjun", createdAt: "2026-06-02T10:00:00.000Z" }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '409':
          description: Already blocked
          content:
            application/json:
              example:
                success: false
                error: { code: CONFLICT, message: "User already blocked" }
        '500': { $ref: '#/components/responses/500InternalError' }
  /api/v1/trust/block/{userId}:
    delete:
      tags: [Trust Center]
      summary: Unblock a user
      operationId: unblockUser
      security:
        - userBearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: User unblocked
          content:
            application/json:
              example:
                success: true
                data: null
                meta: { message: "User unblocked" }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalError' }
...
  /api/v1/trust/report:
    post:
      tags: [Trust Center]
      summary: Report a user for a trust or safety violation
      operationId: reportUser
      security:
        - userBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetUserId, reason]
              properties:
                targetUserId: { type: string, format: uuid }
                reason:       { $ref: '#/components/schemas/FlagReason' }
                description:  { type: string, maxLength: 1000 }
            example:
              targetUserId: "usr-uuid-arjun"
              reason: "FAKE_IDENTITY"
              description: "Profile photos appear to be stock images"
      responses:
        '201':
          description: Report submitted
          content:
            application/json:
              example:
                success: true
                data: { id: "flag-uuid", status: "OPEN", reason: "FAKE_IDENTITY" }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '500': { $ref: '#/components/responses/500InternalError' }
  /api/v1/profile/privacy-controls:
    put:
      tags: [Trust Center]
      summary: Update profile privacy settings
      description: Controls what other users see before a mutual connection is accepted.
      operationId: setPrivacyControls
      security:
        - userBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                showPhotosBeforeMutual:  { type: boolean }
                showBioBeforeMutual:     { type: boolean }
                showAnswersBeforeMutual: { type: boolean }
            example:
              showBioBeforeMutual: false
      responses:
        '200':
          description: Privacy settings updated
          content:
            application/json:
              example:
                success: true
                data:
                  showPhotosBeforeMutual: true
                  showBioBeforeMutual: false
                  showAnswersBeforeMutual: false
                meta: { message: "Privacy settings updated" }
        '400': { $ref: '#/components/responses/400BadRequest' }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalError' }
  /api/v1/profile/pause-visibility:
    post:
      tags: [Trust Center]
      summary: Pause profile visibility (hide from discovery)
      operationId: pauseVisibility
      security:
        - userBearerAuth: []
      responses:
        '200':
          description: Profile hidden from discovery
          content:
            application/json:
              example:
                success: true
                data: { isPaused: true }
                meta: { message: "Profile hidden from discovery" }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalError' }

    delete:
      tags: [Trust Center]
      summary: Resume profile visibility (reappear in discovery)
      operationId: resumeVisibility
      security:
        - userBearerAuth: []
      responses:
        '200':
          description: Profile visible in discovery
          content:
            application/json:
              example:
                success: true
                data: { isPaused: false }
                meta: { message: "Profile visible in discovery" }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '404': { $ref: '#/components/responses/404NotFound' }
        '500': { $ref: '#/components/responses/500InternalError' }
  /api/v1/profile/access-levels:
    get:
      tags: [Trust Center]
      summary: Get the three profile access level definitions
      operationId: getAccessLevels
      security:
        - userBearerAuth: []
      responses:
        '200':
          description: Access level definitions (PUBLIC, TRUSTED, FAMILY)
          content:
            application/json:
              example:
                success: true
                data:
                  - { key: "PUBLIC",  label: "Public",        description: "Visible to everyone in discover.", visibleFields: ["name", "age", "currentCity", "currentCountry", "gender", "verificationStatus"] }
                  - { key: "TRUSTED", label: "Trusted",       description: "Visible after connection accepted.", visibleFields: ["bio", "allPhotos", "realLifeAnswers", "storyPrompts", "voiceIntro", "trustScore"] }
                  - { key: "FAMILY",  label: "Family-aware",  description: "Extended detail when families are involved.", visibleFields: ["contactDetails", "familyBackground", "settlementIntent", "careerDetails"] }
                meta: { total: 3 }
        '401': { $ref: '#/components/responses/401Unauthorized' }
        '500': { $ref: '#/components/responses/500InternalError' }

  # ── Habits ─────────────────────────────────────────────────────────────────
...